web/participant-self-care/pyproject.toml
[tool.poetry]
name = "participant_self_care"
version = "0.1.0"
description = ""
authors = ["Cristian Dinu <c.dinu@ucl.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
authlib = "^1.3.2"
uvicorn = "^0.32.0"
fastapi = "^0.115.2"
itsdangerous = "^2.2.0"
httpx = "^0.27.2"
rich = "^13.9.2"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
aiosqlite = "^0.20.0"
structlog = "^24.4.0"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"
types-toml = "^0.10.8.20240310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
